Changed to separate views from controllers.
Controllers now do not set up their own views. They are configured by
the controller superclass based on the directory where the controller was
loaded from. These views can then be set to something different. The idea
is that the controller itself should not have a dependency on the type of
view used.