Brought back old model of sub-widget initialisation.
So now, when specifying a widget, you create an instance of it, instead
of specifying the class. That way unique constructors can be made, and
the creator can configure the widget instance. The widget template does,
however, come from SpindleWorker, which is given the class of the widget
to find a matching template.